chirpstack-kubernetes/data-parser/requirements.txt
pandas==1.1.3
paho-mqtt==1.5.1
influxdb==5.2.2